backend/requirements.txt
asgiref==3.8.1
certifi==2024.6.2
channels==4.1.0
charset-normalizer==3.3.2
Django==5.0.7
django-cors-headers==4.3.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
flake8==7.0.0
geographiclib==2.0
geopy==2.4.1
gunicorn==22.0.0
idna==3.7
inflection==0.5.1
mccabe==0.7.0
packaging==24.0
pillow==10.3.0
psycopg2-binary==2.9.9
pycodestyle==2.11.1
pyflakes==3.2.0
PyJWT==2.8.0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
requests==2.32.3
setuptools==72.1.0
sqlparse==0.5.0
uritemplate==4.1.1
urllib3==2.2.2
